feat: example client 新增actor管理; dungeonserver 优化, 调通进入游戏流程
feat: 删除 doc, sender 改名

+++ server/internal/protocol/role.go
@@ -32,14 +32,14 @@ type CreateRoleResponse struct {
 type EnterSceneResponse struct {
 	SceneId  uint32    `json:"sceneId"`  // 场景Id
 	RoleInfo *RoleInfo `json:"roleInfo"` // 角色信息
-	PosX     float32   `json:"posX"`     // X坐标
-	PosY     float32   `json:"posY"`     // Y坐标
+	PosX     uint32    `json:"posX"`     // X坐标
+	PosY     uint32    `json:"posY"`     // Y坐标
 }
 
 // MoveRequest 移动请求
 type MoveRequest struct {
-	TargetX float32 `json:"targetX"` // 目标X坐标
-	TargetY float32 `json:"targetY"` // 目标Y坐标
+	TargetX uint32 `json:"targetX"` // 目标X坐标
+	TargetY uint32 `json:"targetY"` // 目标Y坐标
 }
 
 // ErrorResponse 错误响应
